Update response format
